+++ eulerpi/core/data_transformation.py
@@ -1,3 +1,8 @@
+"""Data transformations can be used to improve the performance of the :py:func:`inference <eulerpi.core.inference.inference>` function by improving the quality of the kernel density estimate.
+
+This module contains all predefined data transformations and an abstract base class for custom data transformations.
+"""
+
 from abc import ABC, abstractmethod
 from typing import Optional, Union
 
@@ -7,11 +12,25 @@ from sklearn.decomposition import PCA
 
 
 class DataTransformation(ABC):
-    """Class for normalizing data. The data is normalized by subtracting the mean and multiplying by the inverse of the Cholesky decomposition of the covariance matrix."""
+    """Abstract base class for all data transformations
+
+    Data transformations can be used to improve the performance of the :py:func:`inference <eulerpi.core.inference.inference>` function
+    by improving the quality of the :py:mod:`kernel density estimate <eulerpi.core.kde>`.
+    """
 
     @abstractmethod
     def transform(self, data: jnp.ndarray) -> jnp.ndarray:
-        """Transform the given data."""
+        """Transform the given data point(s)
+
+        Args:
+            data (jnp.ndarray): The data to be transformed. Columns correspond to different dimensions. Rows correspond to different observations.
+
+        Raises:
+            NotImplementedError: Raised if the transform is not implemented in the subclass.
+
+        Returns:
+            jnp.ndarray: The transformed data point(s).
+        """
         raise NotImplementedError()
 
     @abstractmethod