+++ eulerpi/core/inference.py
@@ -1,3 +1,5 @@
+"""The inference module provides the main interface to the eulerpi library in the form of the :py:func:`inference <inference>` function."""
+
 import os
 import pathlib
 from typing import Dict, Optional, Tuple, Union
@@ -40,7 +42,7 @@ def inference(
     Dict[str, np.ndarray],
     ResultManager,
 ]:
-    """Starts the parameter inference for the given model. If a data path is given, it is used to load the data for the model. Else, the default data path of the model is used.
+    """Starts the parameter inference for the given model and data.
 
     Args:
         model(Model): The model describing the mapping from parameters to data.
@@ -59,6 +61,77 @@ def inference(
     Returns:
         Tuple[Dict[str, np.ndarray], Dict[str, np.ndarray], Dict[str, np.ndarray], ResultManager]: The parameter samples, the corresponding simulation results, the corresponding density
         evaluations for each slice and the result manager used for the inference.
+
+    Examples:
+    Inference with eulerpi only requires the model and the data.
+    The following example shows how to run inference for the Covid example model.
+
+    .. code-block:: python
+
+        import numpy as np
+        from eulerpi.examples.corona import Corona
+        from eulerpi.core.inference import inference
+
+        # generate 1000 artificial, 4D data points for the Covid example model
+        data_scales = np.array([1.0, 5.0, 35.0, 2.0])
+        data = (np.random.rand(1000, 4)+1.0)*data_scales
+
+        # run inference only specifying the model and the data
+        (param_sample_dict, sim_res_sample_dict, desities_dict, _) = inference(Corona(), data)
+
+        # retrieve (for instance) the parameter samples by evaluating the parameter sample dictionary in the slice Q0Q1Q2 corresponding to all three joint parameter dimensions
+        param_sample = param_sample_dict["Slice_Q0Q1Q2"]
+
+    Of course, you can also specify additional parameters and import data from a file Data/Coronoa/data.csv.
+    In addition, the result manager is a convenient and versatile interface to access the inference results.
+
+    .. code-block:: python
+
+        import numpy as np
+        from eulerpi.examples.corona import Corona
+        from eulerpi.core.inference import inference
+
+        # run inference with additional arguments
+        (_, _, _, res_manager) = inference(Corona(),
+                                        data = "pathto/data.csv", # load data from a csv file location
+                                        num_processes = 8, # use 8 processes in parallelization
+                                        run_name = "second_run", # specify the run
+                                        num_walkers = 50, # use 50 walkers during sampling
+                                        num_steps = 200, # each walker performs 200 steps
+                                        num_burn_in_samples = 20, # discard the first 20 steps of each walker
+                                        thinning_factor = 10) # only use every 10th sample to avoid autocorrelation
+
+        # use the result manager to retreive the inference results
+        param_sample, sim_res_sample, density_evals = res_manager.load_slice_inference_results(slice = np.array([0,1,2]))
+
+    Principle Compoonent Analysis (PCA) can be useful to reduce the dimension of the data space and is supported by eulerpi.
+    Grid based parameter density estimation is especially useful whenever parameter dimension can be assumed to be indepedent.
+    In the following example, we assume parameter dimension 2 to be independent from dimensions 0 and 1.
+    Therefore, we compute the joint density of dimensions 0 and 1 and the marginal density of dimension 2.
+    Please note, that specifying 30 grid points for a 2D density estimation results in 900 grid points in total.
+    The result manager is especially useful to access already computed inference results and only requires the model and run name.
+
+    .. code-block:: python
+
+        import numpy as np
+        from eulerpi.examples.corona import Corona
+        from eulerpi.core.inference import inference, InferenceType
+        from eulerpi.core.data_transformation_types import DataTransformationType
+        from eulerpi.core.result_manager import ResultManager
+
+        inference(Corona(),
+                data = "pathto/data.csv",
+                slices = [np.array([0,1]), np.array([2])], # specify joint and marginal parameter subdistributions we are interested in
+                inference_type = InferenceType.DENSE_GRID, # use dense grid inference
+                run_name = "grid_run",
+                data_transformation = DataTransformationType.PCA, # perform PCA on the data before inference
+                num_grid_points = 30) # use 30 grid points per parameter dimension
+
+        # initiate the result manager using the model and run name to retreive the inference results computed above and stored offline
+        grid_res_manager = ResultManager(model_name = Corona().name, run_name = "grid_run")
+
+        # load the inference results for the joint distribution of parameter dimensions 0 and 1
+        param_grid_dim01, sim_res_grid_dim01, density_evals_dim01 = grid_res_manager.load_slice_inference_results(slice = np.array([0,1]))
 
     """
 
@@ -91,8 +164,6 @@ def inference(
             "The data_transformation must be one of the enum values of DataTransformationType."
         )
     data = data_transformation.transform(data)
-
-    # TODO rename std_dev to kernel_width, adapt calculation of kernel width
 
     slices = slices or [
         np.arange(model.param_dim)