+++ eulerpi/core/model.py
@@ -1,3 +1,5 @@
+"""Module containing an abstract base class and specialized subclasses for use with the EPI algorithm."""
+
 import inspect
 import os
 import tempfile
@@ -32,7 +34,8 @@ class Model(ABC):
         param_limits(np.ndarray): Box limits for the parameters. The limits are given as a 2D array with shape (param_dim, 2). The parameter limits are used as limits as well as for the movement policy for MCMC sampling, and as boundaries for the grid when using grid-based inference. Overwrite the function param_is_within_domain if the domain is more complex than a box - the grid will still be build based on param_limits, but actual model evaluations only take place within the limits specified in param_is_within_domain. (Default value = None)
         name(str): The name of the model. The class name is used if no name is given. (Default value = None)
 
-    Examples of model implementations can be found in the :doc:`Example Models </examples>`.
+    .. note::
+        Examples of model implementations can be found in the :doc:`Example Models </examples>`.
     """
 
     param_dim: Optional[
@@ -76,6 +79,38 @@ class Model(ABC):
         Returns:
             np.ndarray: The data generated from the parameter.
 
+        Examples:
+        .. code-block:: python
+
+            import numpy as np
+            from eulerpi.examples.heat import Heat
+            from eulerpi.core.model import JaxModel
+            from jax import vmap
+
+            # instantiate the heat model
+            model = Heat()
+
+            # define a 3D example parameter for the heat model
+            example_param = np.array([1.4, 1.6, 0.5])
+
+            # the forward simulation is achieved by using the forward method of the model
+            sim_result = model.forward(example_param)
+
+            # in a more realistic scenario, we would like to perform the forward pass on multiple parameters at once
+            multiple_params = np.array([[1.5, 1.5, 0.5],
+                                        [1.4, 1.4, 0.6],
+                                        [1.6, 1.6, 0.4],
+                                        model.central_param,
+                                        [1.5, 1.4, 0.4]])
+
+            # try to use jax vmap to perform the forward pass on multiple parameters at once
+            if isinstance(model, JaxModel):
+                multiple_sim_results = vmap(model.forward, in_axes=0)(multiple_params)
+
+            # if the model is not a jax model, we can use numpy vectorize to perform the forward pass
+            else:
+                multiple_sim_results = np.vectorize(model.forward, signature="(n)->(m)")(multiple_params)
+
         """
         raise NotImplementedError
 
@@ -89,6 +124,38 @@ class Model(ABC):
         Returns:
             np.ndarray: The jacobian for the variables returned by the :func:`~eulerpi.core.model.Model.forward` method with respect to the parameters.
 
+        Examples:
+        .. code-block:: python
+
+            import numpy as np
+            from eulerpi.examples.heat import Heat
+            from eulerpi.core.model import JaxModel
+            from jax import vmap
+
+            # instantiate the heat model
+            model = Heat()
+
+            # define a 3D example parameter for the heat model
+            example_param = np.array([1.4, 1.6, 0.5])
+
+            sim_jacobian = model.jacobian(example_param)
+
+            # Similar to the forward pass, also the evaluation of the jacobian can be vectorized.
+            # This yields a 3D array of shape (num_params, data_dim, param_dim) = (4,5,3) in this example.
+
+            multiple_params = np.array([[1.5, 1.5, 0.5],
+                                        [1.4, 1.4, 0.6],
+                                        model.central_param,
+                                        [1.5, 1.4, 0.4]])
+
+            # try to use jax vmap for vectorization if possible
+            if isinstance(model, JaxModel):
+                multiple_sim_jacobians = vmap(model.jacobian, in_axes=0)(multiple_params)
+
+            # if the model is not a jax model, we can use numpy vectorize to vectorize
+            else:
+                multiple_sim_jacobians = np.vectorize(model.jacobian, signature="(n)->(m)")(multiple_params)
+
         """
         raise NotImplementedError
 
@@ -97,6 +164,7 @@ class Model(ABC):
     ) -> Tuple[np.ndarray, np.ndarray]:
         """Evaluates the jacobian and the forward pass of the model at the same time. If the method is not overwritten in a subclass it,
         it simply calls :func:`~eulerpi.core.model.Model.forward` and :func:`~eulerpi.core.model.Model.jacobian`.
+        It can be vectorized in the same way as the forward and jacobian methods.
 
         Args:
             param(np.ndarray): The parameter for which the jacobian should be evaluated.
@@ -134,7 +202,6 @@ class Model(ABC):
 class ArtificialModelInterface(ABC):
     """By inheriting from this interface you indicate that you are providing an artificial parameter dataset,
     and the corresponding artificial data dataset, which can be used to compare the results from eulerpi with the ground truth.
-    The comparison can be done using the plotEmceeResults.
 
     """
 
@@ -203,12 +270,16 @@ def add_autodiff(_cls):
 
 
 class JaxModel(Model):
-    """The JaxModel class automatically creates the jacobian method based on the forward method.
-    Additionally it jit compiles the forward and jacobian method with jax.
-    To use this class you have to implement your forward method using jax, e. g. jax.numpy.
-    Dont overwrite the __init__ method of JaxModel without calling the super constructor.
-    Else your forward method wont be jitted.
+    """The JaxModel is a base class for models using the JAX library.
 
+    It automatically creates the jacobian method based on the forward method.
+    Additionally it jit compiles the forward and jacobian method with jax for faster execution.
+
+    .. note::
+
+        To use this class you have to implement your forward method using jax, e. g. jax.numpy.
+        Dont overwrite the __init__ method of JaxModel without calling the super constructor.
+        Else your forward method wont be jitted.
     """
 
     def __init__(
@@ -218,11 +289,6 @@ class JaxModel(Model):
         name: Optional[str] = None,
         **kwargs,
     ) -> None:
-        """Constructor of the JaxModel class.
-
-        Args:
-            name: str: The name of the model. If None the name of the class is used.
-        """
         super().__init__(
             central_param=central_param,
             param_limits=param_limits,
@@ -394,6 +460,7 @@ class SBMLModel(Model):
         self.amici_solver.setAbsoluteTolerance(1e-10)
 
     def setSensitivities(self):
+        """Tell the underlying amici solver to calculate sensitivities based on the attribute `self.param_ids`"""
         if self.param_ids == self.amici_model.getParameterIds():
             self.amici_model.requireSensitivitiesForAllParameters()
         else: